requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
flake8==4.0.1
importlib-metadata==4.11.2
Markdown==3.3.6
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
